Added origin calibration
- Added FloorCalibration class
- Added detection of cross shaped wand
- Add raw data (observations) to Frame
- moved calibration to subdirectory
- added tests for new reconstruction (future releases)

+++ RPIMoCap/RPIMoCapServer/src/Server/floorcalibrationwidget.h
@@ -19,7 +19,7 @@ public:
 
 
 Q_SIGNALS:
-    void snapshotRequest(const double offsetCM);
+    void snapshotRequest(const float offsetCM);
 
 private Q_SLOTS:
     void onSnapshotButton();
